+++ lib/slim_lint/cli.rb
@@ -5,7 +5,7 @@ require 'sysexits'
 
 module SlimLint
   # Command line application interface.
-  class CLI
+  class CLI # rubocop:disable Metrics/ClassLength
     # Create a CLI that outputs to the specified logger.
     #
     # @param logger [SlimLint::Logger]
@@ -38,8 +38,8 @@ module SlimLint
       if options[:help]
         print_help(options)
         Sysexits::EX_OK
-      elsif options[:version]
-        print_version
+      elsif options[:version] || options[:verbose_version]
+        print_version(options)
         Sysexits::EX_OK
       elsif options[:show_linters]
         print_available_linters
@@ -123,8 +123,14 @@ module SlimLint
     end
 
     # Outputs the application name and version.
-    def print_version
+    def print_version(options)
       log.log "#{SlimLint::APP_NAME} #{SlimLint::VERSION}"
+
+      if options[:verbose_version]
+        log.log "slim #{Gem.loaded_specs['slim'].version}"
+        log.log "rubocop #{Gem.loaded_specs['rubocop'].version}"
+        log.log RUBY_DESCRIPTION
+      end
     end
 
     # Outputs the backtrace of an exception with instructions on how to report